pyproject.toml
[project]
name = "streaming-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.54",
    "django>=6.0.2",
    "django-allauth>=65.14.3",
    "django-debug-toolbar>=6.2.0",
    "django-storages>=1.14.6",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
]
